pyproject.toml
[project]
name = "vslpy"
dynamic = [
    "version"
]
description = "Variable Speed of Light Cosmologies"
readme = "README.md"
authors = [
    { name = "Frank Hubeny", email = "frankhubeny@protonmail.com" }
]
maintainers = [
  {name = "Frank Hubeny", email = "frankhubeny@protonmail.com"}
]
license = { text = "BSD-3-Clause" }
requires-python = ">=3.10"
classifiers = [
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Astronomy",
    "Topic :: Scientific/Engineering :: Physics",
]
keywords = [
    "astronomy",
    "astrophysics",
    "cosmology",
    "space",
    "science",
    "units",
    "modeling",
    "models",
    "fitting",
]
dependencies = [
    "numpy>=1.23",
    "pyerfa>=2.0.1.1",
    "astropy-iers-data>=0.2024.7.29.0.32.7",
    "PyYAML>=3.13",
    "packaging>=19.0",
]

[project.urls]
#Homepage = "https://example.com"
#Documentation = "https://readthedocs.org"
Repository = "https://github.com/FrankHubeny/vslpy.git"
"Bug Tracker" = "https://github.com/FrankHubeny/vslpy/issues"
#Changelog = "https://github.com/me/spam/blob/master/CHANGELOG.md"

[build-system]
requires = ["setuptools",
            "setuptools_scm>=6.2"
]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true
license-files = ["LICENSE", "licenses/*.rst"]

[tool.setuptools_scm]
write_to = "vslpy/_version.py"

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = [
    "vslpy",
    "docs",
]
norecursedirs = [
]
vslpy_header = true
doctest_plus = "enabled"
text_file_format = "rst"
remote_data_strict = true
addopts = [
    "--color=yes",
    "--doctest-rst",
    "--strict-config",
    "--strict-markers",
    "-p no:legacypath",
]
log_cli_level = "info"
xfail_strict = true
filterwarnings = [
    "error",
    "ignore:unclosed <socket:ResourceWarning",
    "ignore:unclosed <ssl.SSLSocket:ResourceWarning",
    # Can be removed once Python>=3.11 is the minimum dependency,
    # cf. https://github.com/astropy/astropy/issues/13907
    "ignore:unclosed transport <asyncio.sslproto",
    "ignore:numpy\\.ufunc size changed:RuntimeWarning",
    "ignore:numpy\\.ndarray size changed:RuntimeWarning",
    "ignore:matplotlibrc text\\.usetex:UserWarning:matplotlib",
    # https://github.com/h5py/h5py/pull/2416
    "ignore:__array__ implementation doesn't accept a copy keyword:DeprecationWarning",
]
doctest_norecursedirs = [
]
doctest_subpackage_requires = [
]
markers = [
]

[tool.ruff]
lint.select = ["ALL"]

[project.optional-dependencies]
#gui = ["PyQt5"]
#cli = [
#  "rich",
#  "click",
#]

[project.scripts]
#spam-cli = "spam:main_cli"

[project.gui-scripts]
#spam-gui = "spam:main_gui"

[project.entry-points."spam.magical"]
#tomatoes = "spam:main_tomatoes"
